moveonly: add mp/type-message.h
