+++ test/mp/test/foo-types.h
@@ -9,6 +9,7 @@
 #include <mp/type-context.h>
 #include <mp/type-interface.h>
 #include <mp/type-map.h>
+#include <mp/type-message.h>
 #include <mp/type-number.h>
 #include <mp/type-set.h>
 #include <mp/type-string.h>
